+++ CHANGES.rst
@@ -34,6 +34,8 @@ Changelog
   [chris-adam]
 - Added eSign admin table to manage approvals.
   [chris-adam]
+- Added 'scan_infos' on categorized elements infos.
+  [chris-adam]
 
 3.0 (2021-09-30)
 ----------------